added other CRUD functions (Delete, get all, update,)

+++ src/main/java/com/projectRest/repository/UserRepository.java
@@ -1,11 +1,13 @@
 package com.projectRest.repository;
 
 import com.projectRest.model.User;
+import jakarta.persistence.Id;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
-    boolean  existsById(int id);
+
+    boolean  existsById(long id);
 }
